+++ packages/utils/src/index.ts
@@ -11,4 +11,6 @@ export { errorUtils } from './error_utils';
 export { fetchAsync } from './fetch_async';
 export { signTypedDataUtils } from './sign_typed_data_utils';
 export import AbiEncoder = require('./abi_encoder');
-export { ZeroExTransactionDecoder } from './calldata_decoder';
+export * from './types';
+export { TransactionDecoder } from './transaction_decoder';
+export { ZeroExTransactionDecoder } from './zeroex_transaction_decoder';
