+++ packages/utils/src/transaction_decoder.ts
@@ -1,44 +1,7 @@
-// Decodes any 0x transaction
-
-import * as ContractArtifacts from '@0x/contract-artifacts';
-import { SimpleContractArtifact } from '@0x/types';
-import { AbiDefinition, ContractAbi, MethodAbi } from 'ethereum-types';
+import { AbiDefinition, MethodAbi } from 'ethereum-types';
 import * as _ from 'lodash';
 import { AbiEncoder } from '.';
-import { getContractAddressesForNetworkOrThrow, NetworkId } from '@0x/contract-addresses';
-
-export interface FunctionInfo {
-    functionSignature: string;
-    contractName?: string;
-    contractAddress?: string;
-    networkId?: number;
-    abiEncoder?: AbiEncoder.Method;
-}
-
-interface FunctionInfoBySelector {
-    [index: string]: FunctionInfo[];
-}
-
-export interface DecodedCalldata {
-    functionName: string;
-    functionSignature: string;
-    functionArguments: any;
-}
-
-interface TransactionProperties {
-    contractName?: string;
-    contractAddress?: string;
-    networkId?: number;
-}
-
-interface DeployedContractInfo {
-    contractAddress: string;
-    networkId: number;
-}
-
-interface DeployedContractInfoByName {
-    [index: string]: DeployedContractInfo[];
-}
+import { FunctionInfoBySelector, TransactionData, TransactionProperties, DeployedContractInfo } from './types';
 
 export class TransactionDecoder {
     private _functionInfoBySelector: FunctionInfoBySelector = {};
@@ -86,10 +49,9 @@ export class TransactionDecoder {
         });
     }
 
-    public decode(calldata: string, txProperties_?: TransactionProperties): DecodedCalldata {
+    public decode(calldata: string, txProperties_?: TransactionProperties): TransactionData {
         const functionSelector = TransactionDecoder.getFunctionSelector(calldata);
         const txProperties = _.isUndefined(txProperties_) ? {} : txProperties_;
-        
         const candidateFunctionInfos = this._functionInfoBySelector[functionSelector];
         if (_.isUndefined(candidateFunctionInfos)) {
             throw new Error(`No functions registered for selector '${functionSelector}'`);
@@ -116,50 +78,3 @@ export class TransactionDecoder {
     }
 }
 
-export class ZeroExTransactionDecoder extends TransactionDecoder {
-    private readonly _deployedContractInfoByName = {} as DeployedContractInfoByName;
-    private static _instance: ZeroExTransactionDecoder;
-
-    private static getInstance(): ZeroExTransactionDecoder {
-        if (!ZeroExTransactionDecoder._instance) {
-            ZeroExTransactionDecoder._instance = new ZeroExTransactionDecoder();
-        }
-        return ZeroExTransactionDecoder._instance;
-    }
-
-    private constructor() {
-        super();
-        // Load addresses by contract name
-        _.each(NetworkId, (networkId: any) => {
-            if (typeof networkId !== 'number') return;
-            const networkIdAsNumber = networkId as number;
-            const contractAddressesForNetwork = getContractAddressesForNetworkOrThrow(networkIdAsNumber);
-            _.each(contractAddressesForNetwork, (contractAddress: string, contractName: string) => {
-                const contractNameLowercase = _.toLower(contractName);
-                if (_.isUndefined(this._deployedContractInfoByName[contractNameLowercase])) {
-                    this._deployedContractInfoByName[contractNameLowercase] = [];
-                }
-                this._deployedContractInfoByName[contractNameLowercase].push({contractAddress, networkId: networkIdAsNumber});
-            });
-        });
-        // Load contract artifacts
-        _.each(ContractArtifacts, (contractArtifactAsJson: any) => {
-            const conractArtifact = contractArtifactAsJson as SimpleContractArtifact;
-            const contractName = conractArtifact.contractName;
-            const contractNameLowercase = _.toLower(contractName);
-            const contractAbi: ContractAbi = conractArtifact.compilerOutput.abi;
-             this.addABI(contractAbi, contractName, this._deployedContractInfoByName[contractNameLowercase]);
-        });
-    }
-
-    public static addABI(abiArray: AbiDefinition[], contractName: string, deploymentInfos?: DeployedContractInfo[]): void {
-        const instance = ZeroExTransactionDecoder.getInstance();
-        instance.addABI(abiArray, contractName, deploymentInfos);
-    }
-
-    public static decode(calldata: string, txProperties?: TransactionProperties): DecodedCalldata {   
-        const instance = ZeroExTransactionDecoder.getInstance();
-        const decodedCalldata = instance.decode(calldata, txProperties);
-        return decodedCalldata;
-    }
-}
