+++ packages/utils/src/types.ts
@@ -1,0 +1,34 @@
+import { AbiEncoder } from '.';
+
+export interface FunctionInfo {
+    functionSignature: string;
+    contractName?: string;
+    contractAddress?: string;
+    networkId?: number;
+    abiEncoder?: AbiEncoder.Method;
+}
+
+export interface FunctionInfoBySelector {
+    [index: string]: FunctionInfo[];
+}
+
+export interface TransactionData {
+    functionName: string;
+    functionSignature: string;
+    functionArguments: any;
+}
+
+export interface TransactionProperties {
+    contractName?: string;
+    contractAddress?: string;
+    networkId?: number;
+}
+
+export interface DeployedContractInfo {
+    contractAddress: string;
+    networkId: number;
+}
+
+export interface DeployedContractInfoByName {
+    [index: string]: DeployedContractInfo[];
+}
